% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBManagementConfig.R
\name{BuildDBEnvirontment}
\alias{BuildDBEnvirontment}
\title{Build DB Environment.}
\usage{
BuildDBEnvirontment(platform.c, slave.flag = FALSE, pool.flag = FALSE)
}
\arguments{
\item{platform.c}{Database type.}

\item{slave.flag}{TRUE/FALSE flag to determinate if it is a queries slave server.}

\item{pool.flag}{TRUE/FALSE flag to determinate if it is a pool connection.}
}
\value{
None
}
\description{
Method to generate all basic information to connect into server.
}
