% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetConnections.R
\name{ClosePullConnectionDB}
\alias{ClosePullConnectionDB}
\title{Close Pull Connections}
\usage{
ClosePullConnectionDB()
}
\value{
None
}
\description{
This function closes a pool of mysql connections.
}
