% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBManagementConfig.R
\name{DBSwitchTechnology}
\alias{DBSwitchTechnology}
\title{Get a database connection}
\usage{
DBSwitchTechnology(server.config, pool.flag)
}
\arguments{
\item{server.config}{Server configuration.}

\item{pool.flag}{TRUE/FALSE flag to determine if it is a pool connection.}
}
\value{
None
}
\description{
Get a connection depending on driver technology.
}
