% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBManagementConfig.R
\name{GenerateSchemaList}
\alias{GenerateSchemaList}
\title{Build schema info into environtment variables}
\usage{
GenerateSchemaList(db.servers.list, pool.flag)
}
\arguments{
\item{db.servers.list}{List of all availables servers.}

\item{pool.flag}{TRUE/FALSE flag to determine if it is a pool connection.}
}
\value{
None
}
\description{
Assign availables Schemas into environment variables.
}
