% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetConnections.R
\name{GetMySQLConnection}
\alias{GetMySQLConnection}
\title{Get MySQL connection}
\usage{
GetMySQLConnection(
  db.host,
  db.user,
  db.pass,
  db.port,
  db.default,
  db.compression = 32,
  db.pool = FALSE
)
}
\arguments{
\item{db.host}{string identifying the host machine running the MariaDB server.}

\item{db.user}{Username}

\item{db.pass}{Password .}

\item{db.port}{Integer of the tcp/ip port to establish a connection.}

\item{db.compression}{Default: 32.}

\item{db.pool}{Pool flag. By Default: False. If it is true, function generates a pool connection to mysql server.}
}
\value{
None
}
\description{
This function establishes a mysql connection.
}
