% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBManagementConfig.R
\name{GetSchemaInfo}
\alias{GetSchemaInfo}
\title{Get Schema info}
\usage{
GetSchemaInfo(connection.method, pool.flag)
}
\arguments{
\item{connection.method}{a MariaDBConnection object, produced by DBI::dbConnect().}

\item{pool.flag}{TRUE/FALSE flag to determine if it is a pool connection.}
}
\value{
None
}
\description{
This function generates a list of availables schemas from an estabished connection.
}
